backend/requirements.txt
cohere
torch
torchvision
numpy
pillow
pdf2image